+++ src/surrealist/__init__.py
@@ -3,9 +3,10 @@ from .errors import *
 from .ql import Database, DatabaseConnectionsPool, Table, Where
 from .result import SurrealResult
 from .surreal import Surreal
-from .utils import DATA_LENGTH_FOR_LOGS, get_uuid
+from .utils import DATA_LENGTH_FOR_LOGS, get_uuid, to_surreal_datetime_str, to_datetime
 
 __all__ = ("Surreal", "SurrealResult", "WebSocketConnection", "HttpConnection", "PySurrealError", "HttpConnectionError",
            "HttpClientError", "SurrealConnectionError", "WebSocketConnectionError", "WebSocketConnectionClosedError",
            "ConnectionParametersError", "CompatibilityError", "OperationOnClosedConnectionError", "WrongCallError",
-           "DATA_LENGTH_FOR_LOGS", "Connection", "get_uuid", "Database", "Table", "Where", "DatabaseConnectionsPool")
+           "DATA_LENGTH_FOR_LOGS", "Connection", "get_uuid", "Database", "Table", "Where", "DatabaseConnectionsPool",
+           "to_surreal_datetime_str", "to_datetime")