+++ src/surrealist/ql/statements/show.py
@@ -1,8 +1,8 @@
-from datetime import datetime
+from datetime import datetime, timezone
 from typing import List, Optional
 
 from surrealist.connections import Connection
-from surrealist.utils import OK, DATE_FORMAT, DATE_FORMAT_NS
+from surrealist.utils import OK, DATE_FORMAT, DATE_FORMAT_NS, to_surreal_datetime_str
 from .statement import Statement
 
 
@@ -28,7 +28,7 @@ class Show(Statement):
                 format_ = DATE_FORMAT if "." not in self._since else DATE_FORMAT_NS
                 datetime.strptime(self._since, format_)
             except ValueError:
-                result.append("Timestamp in wrong format, you need iso-date like 2024-01-01T10:10:10.000001Z")
+                result.append("Timestamp in the wrong format, you need iso-date like 2024-01-01T10:10:10.000001Z")
         if self._limit and self._limit < 1:
             result.append("Limit should not be less than 1")
         return [OK] if not result else result
@@ -43,6 +43,6 @@ class Show(Statement):
 
     def _clean_str(self):
         if not self._since:
-            self._since = f'{datetime.utcnow().isoformat("T")}Z'  # default value
+            self._since = to_surreal_datetime_str(datetime.now(timezone.utc))  # default value
         limit = f" LIMIT {self._limit}" if self._limit else ""
         return f'SHOW CHANGES FOR TABLE {self._table_name} SINCE "{self._since}"{limit}'